+++ Extensions.Constants/VersionHistory.md
@@ -1,5 +1,8 @@
-﻿# Extensions.Constants Version History.
+﻿# GCCHigh.Extensions.Constants Version History.
 
 ### **6.0.800 (2024-02-27)**<br>
 	- Rearchitected and Optimized.<br>
 	- Rebranded the `Extensions` set of NuGet packages as `GCCHigh.Extensions`.
+
+### **6.1.800 (2024-02-27)**<br>
+	- Added `UserMembershipType` type to Constants.