+++ Extensions.Core/VersionHistory.md
@@ -21,3 +21,6 @@
 		
 ### **6.5.800 (2024-02-27)**<br>
 	- Optimized tenant initialization.
+		
+### **6.6.800 (2024-02-27)**<br>
+	- Optimized tenant initialization.