+++ Extensions.Logit/VersionHistory.md
@@ -3,6 +3,9 @@
 ### **4.8.700 (2023-03-24)**<br>
 	- Separate `System.Logit` class for logging functionality from Extensions.cs.
 
-### **6.0.800 (2024-02-27)**<br>
+### **6.0.800 (2024-02-26)**<br>
 	- Rearchitected and Optimized.<br>
 	- Rebranded the `Extensions` set of NuGet packages as `GCCHigh.Extensions`.
+
+### **6.1.800 (2024-02-27)**<br>
+	- NuGet dependencies.<br>